Integrate prefix-cache benchmark into v2 CommandFactory entry + thin launcher
Rename tt-inference-server-v2/workflow_runner.py to run.py as the single v2
entry point and wire the prefix-caching benchmark through the CommandFactory
architecture: prefix-cache CLI flags live in run.py's parser, and
command_factory.py builds PrefixCacheOptions (plus the debug-test JWT) into
OrchestratorMetadata, which BenchmarksWorkflow already consumes.

Keep run.py free of import-time side effects by moving venv selection into a
thin run_prefix_cache.py launcher that ensures the V2_PREFIX_CACHE venv and
exec's run.py once -- the same external-venv-selection pattern v2_bridge.py
uses for image models. Point v2_bridge.py at run.py and refresh the README.
